+++ src/vidore_benchmark/ocr/__init__.py
@@ -1,0 +1,2 @@
+from .interfaces import BoundingBox, ExtractedWord
+from .tesseract_utils import to_extracted_words