+++ src/vidore_benchmark/retrievers/__init__.py
@@ -2,34 +2,57 @@ try:
     from .bge_m3_retriever import BGEM3Retriever
 except ImportError:
     pass
+
 try:
     from .bge_m3_colbert_retriever import BGEM3ColbertRetriever
 except ImportError:
     pass
+
 try:
     from .bm25_retriever import BM25Retriever
 except ImportError:
     pass
+
+try:
+    from .cohere_api_retriever import CohereAPIRetriever
+except ImportError:
+    pass
+
 try:
     from .colpali_retriever import ColPaliRetriever
 except ImportError:
     pass
+
+try:
+    from .colqwen_retriever import ColQwenRetriever
+except ImportError:
+    pass
+
+try:
+    from .dse_retriever import DSERetriever
+except ImportError:
+    pass
+
 try:
     from .dummy_retriever import DummyRetriever
 except ImportError:
     pass
+
 try:
     from .jina_clip_retriever import JinaClipRetriever
 except ImportError:
     pass
+
 try:
     from .nomic_retriever import NomicVisionRetriever
 except ImportError:
     pass
+
 try:
     from .siglip_retriever import SigLIPRetriever
 except ImportError:
     pass
+
 try:
     from .vision_retriever import VisionRetriever
 except ImportError: