+++ src/vidore_benchmark/retrievers/jina_clip_retriever.py
@@ -4,6 +4,7 @@ import math
 from typing import List, Optional, cast
 
 import torch
+from colpali_engine.utils.torch_utils import get_torch_device
 from PIL import Image
 from tqdm import tqdm
 from transformers import AutoModel
@@ -11,15 +12,23 @@ from transformers import AutoModel
 from vidore_benchmark.retrievers.utils.register_retriever import register_vision_retriever
 from vidore_benchmark.retrievers.vision_retriever import VisionRetriever
 from vidore_benchmark.utils.iter_utils import batched
-from vidore_benchmark.utils.torch_utils import get_torch_device
 
 
-@register_vision_retriever("jinaai/jina-clip-v1")
+@register_vision_retriever("jina-clip-v1")
 class JinaClipRetriever(VisionRetriever):
-    def __init__(self, device: str = "auto"):
+    def __init__(
+        self,
+        pretrained_model_name_or_path: str = "jinaai/jina-clip-v1",
+        device: str = "auto",
+    ):
         super().__init__()
+        self.pretrained_model_name_or_path = pretrained_model_name_or_path
         self.device = get_torch_device(device)
-        self.model = AutoModel.from_pretrained("jinaai/jina-clip-v1", trust_remote_code=True).to(self.device).eval()
+
+        self.model = (
+            AutoModel.from_pretrained(self.pretrained_model_name_or_path, trust_remote_code=True).to(self.device).eval()
+        )
+
         self.emb_dim_query = 768
         self.emb_dim_doc = 768