+++ tests/retrievers/test_bge_m3_retriever.py
@@ -13,9 +13,9 @@ def retriever() -> Generator[BGEM3Retriever, None, None]:
 
 
 @pytest.mark.slow
-def test_forward_queries(retriever: BGEM3Retriever, queries_fixtures):
-    embeddings_queries = retriever.forward_queries(queries_fixtures, batch_size=1)
-    assert len(embeddings_queries) == len(queries_fixtures)
+def test_forward_queries(retriever: BGEM3Retriever, queries_fixture):
+    embeddings_queries = retriever.forward_queries(queries_fixture, batch_size=1)
+    assert len(embeddings_queries) == len(queries_fixture)
 
 
 @pytest.mark.slow
@@ -25,8 +25,8 @@ def test_forward_documents(retriever: BGEM3Retriever, document_ocr_text_fixture)
 
 
 @pytest.mark.slow
-def test_get_scores(retriever: BGEM3Retriever, queries_fixtures, document_ocr_text_fixture):
-    emb_query = retriever.forward_queries(queries_fixtures, batch_size=1)
+def test_get_scores(retriever: BGEM3Retriever, queries_fixture, document_ocr_text_fixture):
+    emb_query = retriever.forward_queries(queries_fixture, batch_size=1)
     emb_doc = retriever.forward_documents(document_ocr_text_fixture, batch_size=1)
     scores = retriever.get_scores(emb_query, emb_doc)
-    assert scores.shape == (len(queries_fixtures), len(document_ocr_text_fixture))
+    assert scores.shape == (len(queries_fixture), len(document_ocr_text_fixture))